Update to current yogs.

+++ code/game/objects/effects/decals/cleanable.dm
@@ -30,21 +30,21 @@
 		return TRUE
 
 /obj/effect/decal/cleanable/attackby(obj/item/W, mob/user, params)
-	if(istype(W, /obj/item/reagent_containers/glass) || istype(W, /obj/item/reagent_containers/food/drinks))
+	if((istype(W, /obj/item/reagent_containers/glass) && !istype(W, /obj/item/reagent_containers/glass/rag)) || istype(W, /obj/item/reagent_containers/food/drinks))
 		if(src.reagents && W.reagents)
 			. = 1 //so the containers don't splash their content on the src while scooping.
 			if(!src.reagents.total_volume)
-				to_chat(user, "<span class='notice'>[src] isn't thick enough to scoop up!</span>")
+				to_chat(user, span_notice("[src] isn't thick enough to scoop up!"))
 				return
 			if(W.reagents.total_volume >= W.reagents.maximum_volume)
-				to_chat(user, "<span class='notice'>[W] is full!</span>")
+				to_chat(user, span_notice("[W] is full!"))
 				return
-			to_chat(user, "<span class='notice'>You start scooping up [src] into [W]...</span>")
+			to_chat(user, span_notice("You start scooping up [src] into [W]..."))
 			var/scoop_time
 			scoop_time = min((W.reagents.maximum_volume - W.reagents.total_volume), src.reagents.total_volume) //1 second per 10 units scooped
 			if(do_mob(user, user, scoop_time))
 				if(src)
-					to_chat(user, "<span class='notice'>You scoop up [src] into [W]!</span>")
+					to_chat(user, span_notice("You scoop up [src] into [W]!"))
 					reagents.trans_to(W, reagents.total_volume, transfered_by = user)
 					if(!reagents.total_volume) //scooped up all of it
 						qdel(src)
@@ -56,7 +56,7 @@
 		else
 			var/hotness = W.is_hot()
 			reagents.expose_temperature(hotness)
-			to_chat(user, "<span class='notice'>You heat [name] with [W]!</span>")
+			to_chat(user, span_notice("You heat [name] with [W]!"))
 	else
 		return ..()
 
@@ -99,3 +99,8 @@
 		return bloodiness
 	else
 		return 0
+
+/obj/effect/decal/cleanable/wash(clean_types)
+	..()
+	qdel(src)
+	return TRUE
